policies/requirements.txt
Django==2.2.3
djangorestframework==3.9.4
gunicorn==19.9.0
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
